Remove old symbol naming code.
